+++ artiq/frontend/artiq_dashboard.py
@@ -13,7 +13,7 @@ from qasync import QEventLoop
 from sipyco.pc_rpc import AsyncioClient, Client
 from sipyco.broadcast import Receiver
 from sipyco import common_args
-from sipyco.tools import atexit_register_coroutine
+from sipyco.tools import atexit_register_coroutine, SimpleSSLConfig
 from sipyco.sync_struct import Subscriber
 
 from artiq import __artiq_dir__ as artiq_dir, __version__ as artiq_version
@@ -42,6 +42,13 @@ def get_argparser():
     parser.add_argument(
         "--port-broadcast", default=1067, type=int,
         help="TCP port to connect to for broadcasts (default: %(default)s)")
+    parser.add_argument(
+        "--ssl", nargs=3, metavar=('CERT', 'KEY', 'PEER'), default=None,
+        help="Enable SSL authentication: "
+             "CERT: client certificate file, "
+             "KEY: client private key, "
+             "PEER: server certificate to trust "
+             "(default: %(default)s)")
     parser.add_argument(
         "--db-file", default=None,
         help="database file for local GUI settings (default: %(default)s)")
@@ -133,6 +140,7 @@ def main():
                                     server=args.server.replace(":", "."),
                                     port=args.port_notify))
 
+    ssl_config = SimpleSSLConfig(*args.ssl) if args.ssl else None
     forced_platform = []
     if (QtGui.QGuiApplication.platformName() == "wayland" and
             not os.getenv("QT_QPA_PLATFORM")):
@@ -149,11 +157,11 @@ def main():
     for target in "schedule", "experiment_db", "dataset_db", "device_db", "interactive_arg_db":
         client = AsyncioClient()
         loop.run_until_complete(client.connect_rpc(
-            args.server, args.port_control, target))
+            args.server, args.port_control, target, ssl_config=ssl_config))
         atexit_register_coroutine(client.close_rpc)
         rpc_clients[target] = client
 
-    master_management = Client(args.server, args.port_control, "master_management")
+    master_management = Client(args.server, args.port_control, "master_management", ssl_config=ssl_config)
     try:
         server_name = master_management.get_name()
     finally:
@@ -176,7 +184,7 @@ def main():
                                   ("interactive_args", interactive_args.Model)):
         subscriber = ModelSubscriber(notifier_name, modelf, report_disconnect)
         loop.run_until_complete(subscriber.connect(
-            args.server, args.port_notify))
+            args.server, args.port_notify, ssl_config=ssl_config))
         atexit_register_coroutine(subscriber.close, loop=loop)
         sub_clients[notifier_name] = subscriber
 
@@ -184,7 +192,7 @@ def main():
     for target in "log", "ccb":
         client = Receiver(target, [], report_disconnect)
         loop.run_until_complete(client.connect(
-            args.server, args.port_broadcast))
+            args.server, args.port_broadcast, ssl_config=ssl_config))
         atexit_register_coroutine(client.close, loop=loop)
         broadcast_clients[target] = client
 
@@ -215,7 +223,8 @@ def main():
     smgr.register(d_explorer)
 
     d_datasets = datasets.DatasetsDock(sub_clients["datasets"],
-                                       rpc_clients["dataset_db"])
+                                       rpc_clients["dataset_db"],
+                                       loop)
     smgr.register(d_datasets)
 
     d_applets = applets_ccb.AppletsCCBDock(main_window,
@@ -247,6 +256,7 @@ def main():
         sub_clients["interactive_args"],
         rpc_clients["interactive_arg_db"]
     )
+    smgr.register(d_interactive_args)
 
     d_schedule = schedule.ScheduleDock(
         rpc_clients["schedule"], sub_clients["schedule"])
@@ -269,11 +279,6 @@ def main():
     main_window.addDockWidget(QtCore.Qt.DockWidgetArea.BottomDockWidgetArea, d_schedule)
 
     # load/initialize state
-    if os.name == "nt":
-        # HACK: show the main window before creating applets.
-        # Otherwise, the windows of those applets that are in detached
-        # QDockWidgets fail to be embedded.
-        main_window.show()
     smgr.load()
 
     def init_cbs(ddb):
@@ -281,7 +286,7 @@ def main():
         d_waveform.init_ddb(ddb)
         return ddb
     devices_sub = Subscriber("devices", init_cbs, [d_ttl_dds.dm.notify_ddb, d_waveform.notify_ddb])
-    loop.run_until_complete(devices_sub.connect(args.server, args.port_notify))
+    loop.run_until_complete(devices_sub.connect(args.server, args.port_notify, ssl_config=ssl_config))
     atexit_register_coroutine(devices_sub.close, loop=loop)
 
     smgr.start(loop=loop)