+++ RELEASE_NOTES.rst
@@ -9,6 +9,7 @@ ARTIQ-9 (Unreleased)
 * Hardware support:
    - 12Gbps CoaXPress grabber support on Kasli-SoC with CoaXPress-SFP adapter
      (and ZC706 with Hello-FPGA CXP 4R FMC card).
+   - Songbird - a 4-tone, 16-bit DDS using a 2.5Gsps LTC2000 DAC.
    - Improved SDRAM memory controller and DMA cores puts Kasli DMA performance on par with
      other platforms.
    - Core device reflashing over the network through the new ``flash`` tool in ``artiq_coremgmt``.
@@ -21,6 +22,7 @@ ARTIQ-9 (Unreleased)
    - Fastino monitoring with Moninj.
    - Zotino monitoring now displays the values in volts.
    - artiq_flash can now flash Phaser through a Digilent HS2 Programming cable.
+   - artiq_flash can now erase specified flash region(s). 
    - Support for the ultra-low-cost EBAZ4205 Zynq-7000 control card, with core device driver
      for the AD9834 DDS, tested with the ZonRi Technology Co., Ltd. AD9834-Module.
    - Configurable number of Grabber ROI engines through the ``roi_engine_count``
@@ -34,6 +36,9 @@ ARTIQ-9 (Unreleased)
      header context menu.
    - State files are now automatically backed up upon successful loading.
    - Experiments can now use ``restart_applet`` CCB command to restart applets.
+* ``artiq_session`` has been revamped. Output now indicates the originating process. Dashboard
+   is started after the controller manager. All processes are cleanly terminated after dashboard
+   is closed.
 * ``afws_client`` now uses the "happy eyeballs" algorithm (RFC 6555) for a faster and more
   reliable connection to the server.
 * Compiler can now give automatic suggestions for ``kernel_invariants``. 
@@ -41,6 +46,7 @@ ARTIQ-9 (Unreleased)
 * Python 3.12 and 3.13 support.
 * The Zadig driver installer was added to the MSYS2 offline installer.
 * ``artiq.coredevice.fmcdio_vhdci_eem`` has been removed.
+* Optional SSL support with mutual certificate authentication was added between components running on computers.
 
 Breaking changes:
 
@@ -53,6 +59,7 @@ Breaking changes:
      require PYON v2: ensure sipyco v2 is used.
    - When PYON v2 decoding fails in the following contexts, decoding as PYON v1 is attempted.
      The support for PYON v1 decoding will be removed in a future ARTIQ release.
+
       + Dataset DB file
       + HDF5 results ``expid``. Note that existing files are never altered.
       + artiq_client set-dataset
@@ -61,6 +68,19 @@ Breaking changes:
       + Interactive arguments (artiq_run, artiq_client)
       + Waveform channel list file
       + devarg overrides
+* ``artiq_coremgmt config`` can now read multiple times in the same command and read binary files
+  from the core device config to host. Instead of ``read KEY``, use ``read -s KEY``.
+* The variant for EFC and KC705 gateware builds must always be specified.
+* ``artiq_flash``: the syntax is different when operating on partial regions of the flash. See the example below.
+  The new syntax also allows erasing partial regions. Backward compatibility is retained when erasing and 
+  flashing all regions.
+
+::
+
+  # Erase the storage flash region, flash the gateware and firmware but not bootloader and then restart the FPGA device
+  artiq_flash erase=storage write=gateware,firmware start -d ./artiq_kasli/master/ --srcbuild
+
+* Experimental features have been removed.
 
 ARTIQ-8
 -------